RN-404: Moved SqlQuery to @tupaia/database (#3860)
