+++ packages/database/src/index.js
@@ -20,3 +20,4 @@ export { DatabaseChangeChannel } from './DatabaseChangeChannel';
 export { DatabaseModel } from './DatabaseModel';
 export { DatabaseType } from './DatabaseType';
 export * from './testUtilities';
+export { SqlQuery } from './SqlQuery';
